requirements.txt
# WebRTC Signaling Server Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Management
python-dotenv==1.0.0

# WebSocket Support (included with FastAPI)
websockets==12.0

# Logging and Monitoring
structlog==23.2.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For production deployment
gunicorn==21.2.0

# Security Dependencies
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
cryptography>=40.0.0

# SSL/TLS Support
pyOpenSSL>=23.0.0
